pyproject.toml
[project]
name = "content-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-generativeai>=0.8.6",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.54.0",
]
